Revert "window-actor: Use int for opacity"
This reverts commit 8448e692504b5e71dc4d66b36ca68da7afa9a42b.

+++ src/compositor/meta-shadow-factory.c
@@ -203,7 +203,7 @@ meta_shadow_paint (MetaShadow     *shadow,
                    int             window_y,
                    int             window_width,
                    int             window_height,
-                   int             opacity,
+                   guint8          opacity,
                    cairo_region_t *clip,
                    gboolean        clip_strictly)
 {
